login: add 'isUserAdmin' to session data
