+++ src/teuthology_api/routes/login.py
@@ -5,6 +5,8 @@ from fastapi.responses import RedirectResponse
 from dotenv import load_dotenv
 import httpx
 
+from teuthology_api.services.helpers import isAdmin
+
 load_dotenv()
 
 GH_CLIENT_ID = os.getenv("GH_CLIENT_ID")
@@ -85,9 +87,12 @@ async def handle_callback(code: str, request: Request):
             "access_token": token,
         }
         request.session["user"] = data
+        isUserAdmin = await isAdmin(data["username"], data["access_token"])
+        data["isUserAdmin"] = isUserAdmin
     cookie_data = {
         "username": data["username"],
         "avatar_url": response_org_dic.get("user", {}).get("avatar_url"),
+        "isUserAdmin": isUserAdmin,
     }
     cookie = "; ".join(
         [f"{str(key)}={str(value)}" for key, value in cookie_data.items()]
